fixed float bug

+++ candidbot_controller.py
@@ -52,14 +52,14 @@ class CandidbotController:
 
     if debug_display:
       # Draw navigation regions
-      center_x = int(self.frame_width / 2)
+      center_x = self.frame_width / 2
       photo_region = (
         (0, self.frame_height - self.dist_thresh), (self.frame_width, self.frame_height))
       center_region = (
         (center_x - self.center_thresh, 0), (center_x + self.center_thresh, self.frame_height))
 
-      draw_region(frame, photo_region[0], photo_region[1], (0, 0, 255))
-      draw_region(frame, center_region[0], center_region[1], (0, 255, 0))
+      draw_region(frame, int(photo_region[0]), int(photo_region[1]), (0, 0, 255))
+      draw_region(frame, int(center_region[0]), int(center_region[1]), (0, 255, 0))
 
       # Draw green circle on target
       center = self.tracker.get_centroid(self.target.box)
